Initial Commit
Initial Commit of the AD Module. THe AD Module is a combination hardware
module that combines the functionality of the analog sensor module the
digitial sensor module and the module hub into a single piece of
hardware that will stream sensor data and other info to a wireless
gateway to the Vivaplanet cloud.